requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-generativeai==0.8.2
google-api-python-client==2.147.0
redis==5.0.4
celery==5.4.0
django-celery-results==2.5.1
youtube-transcript-api==0.6.2